Fix WindiCSS preprocessing errors
The idea is to add a svelte.config.js file that instructs svelte (and the svelte extension in vscode) to preprocess windicss. This however inflicted changes in the way that the css was processed and caused inconsistencies with before and after adding `svelte-windicss-preprocess` - the fix was to add `important: true` to the windicss config.

To lint css, I added stylelint, which also defines the code format with tabs and etc.

The UI has been changed a bit to better match how the flow will be (conditional classes, `@apply`, external css, etc).

An issue with `@apply` is that it can't apply to conditionally applied classes, so conditional classes will need to be added inline, they can't be added via @applied styles.