Implement GraphQL's urql library and tweak project for GraphQL
The chat page was overhauled to use material components.

Automation was setup to fetch the schema of a dev server which allows to generate types for usage in code.

Finally, the getApiUrl function of utils was reworked to use the URL object,  which should be backward compatible while allowing more flexibility in the return object.