Update chatting behavior to show message in list as inactive then active
When initially sending the message, it will show up in the list, with no time appended and greyed out - this means it hasn't been sent / you haven't received the notification that it was successfully sent yet. Once the confirmation is received, you get the full active line with a date and proper text.